Add failing py3 tests on PUB
Fix where examples do the wrong thing in Py3
