+++ example/make_requests.py
@@ -27,7 +27,7 @@ client_id = ''.join(
 def sid_on_msg(nats_protocol, sid, subject, reply_to, payload):
     stdout.write("sid: {}, subject: {}, reply-to: {}\r\n".format(
         sid, subject, reply_to))
-    stdout.write(payload)
+    stdout.write(payload.decode())
     stdout.write("\r\n*")
