+++ example/queue_respond.py
@@ -25,7 +25,7 @@ def respond_on_msg(nats_protocol, sid, subject, reply_to, payload):
     Write the message payload to standard out, and if
     there is a reply_to, publish a message to it.
     """
-    stdout.write(payload)
+    stdout.write(payload.decode())
     stdout.write("\r\n")
     if reply_to:
         nats_protocol.pub(reply_to, "Roger, from {}!".format(responder_id))
