+++ txnats/_meta.py
@@ -1,2 +1,2 @@
-version_info = (0, 5, 0)
+version_info = (0, 5, 1)
 version = '.'.join(map(str, version_info))
